% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenLaplace.R
\name{eigenLaplace}
\alias{eigenLaplace}
\title{Generate eigenvalues of discrete Laplace matrix.}
\usage{
eigenLaplace(mm, nn)
}
\arguments{
\item{mm}{Number of rows of the discrete Laplace matrix.}

\item{nn}{Number of columns of the discrete Laplace matrix.}
}
\value{
A row vector containing the eigenvalues of the discrete
    laplace matrix.
}
\description{
The eigenvalues of a discrete Laplace matrix with dimension (\code{mm}, \code{nn}) are
calculated.
}
\examples{
eigval <- eigenLaplace(5, 5)

}
