% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpFunctions.R
\name{turnmat}
\alias{turnmat}
\title{Turn matrix 90 degrees counter-clockwise.}
\usage{
turnmat(x)
}
\arguments{
\item{x}{Matrix to be turned.}
}
\value{
Matrix \code{x}, turned by 90 degrees counter-clockwise.
}
\description{
Help function to turn matrix 90 degrees counter-clockwise.
\code{turnmat} is used as an internal function in some of the plotting 
functions. Depending on how the data is stored, it can be necessary to turn 
the matrices 90 degrees counter-clockwise for being able to plot them correctly.
}
\examples{
set.seed(987)
sampleMat <- matrix(stats::rnorm(100), nrow = 10)
sampleMatTurn <- turnmat(x = sampleMat)

}
